requirements.txt
Django==4.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
flake8==6.1.0
black==23.11.0
isort==5.12.0
